personal file area: reduce memory load on folders with many many files by implementing a pagination
this changeset/bugfix is a bit large due to the fact

a) that the pagination should be intuitive, i.e. if the users changes back to an upper directory, the former directory should be visible on the current page. same goes for actions such as editing...
b) the pagination should be configurable. now the user may choose whether he wants to see 20, 50 or 100 entries on one page. this value is stored in the user config
c) the personal file area is kinda messed up and i couldn't help myself and tweaked a few things here and there

all in all this shouldn't be transfered to any previous branch since the fatal error only occurs in edge cases (folders with way more than 2k files). but if it should be transfered to another branch nevertheless, let me know and i will do so.

fixes #5088